move unit specs to test/unit
