+++ test/unit/geometry/cone.spec.js
@@ -1,10 +1,10 @@
-require('../common/');
+require('../../common/');
 const assert = require('assert');
-const qtek = require('../../dist/qtek');
+const qtek = require('../../../dist/qtek');
 
-describe('Cylinder.Spec', function () {
+describe('Cone.Spec', function () {
     it('constructor', function () {
-        const geometry = new qtek.geometry.Cylinder({
+        const geometry = new qtek.geometry.Cone({
             widthSegments : 1,
             heightSegments : 2,
             depthSegments : 3
@@ -19,40 +19,40 @@ describe('Cylinder.Spec', function () {
     });
 
     it('#generateFaceNormals', function () {
-        const geometry = new qtek.geometry.Cylinder();
+        const geometry = new qtek.geometry.Cone();
         const normal = geometry.attributes.normal.value;
         geometry.generateFaceNormals();
         assert.notDeepEqual(geometry.attributes.normal.value, normal);
     });
     
     it('#generateTangents', function () {
-        const geometry = new qtek.geometry.Cylinder();
+        const geometry = new qtek.geometry.Cone();
         const tangent = geometry.attributes.tangent.value;
         geometry.generateTangents();
         assert.notDeepEqual(geometry.attributes.tangent.value, tangent);
         const len = geometry.attributes.tangent.value.length;
-        assert(len === 808, 'Cylinder.tangent : ' + len);
+        assert(len === 328, 'Cone.tangent : ' + len);
     });
 
     it('#generateUniqueVertex', function () {
-        const geometry = new qtek.geometry.Cylinder();
+        const geometry = new qtek.geometry.Cone();
         assert(!geometry.isUniqueVertex());
         let normals = geometry.attributes.normal.value,
             texCoords = geometry.attributes.texcoord0.value;
-        assert(normals.length === 606, 'Cylinder.normal.count : ' + normals.length);
-        assert(texCoords.length === 404, 'Cylinder.textcoord.count : ' + texCoords.length);
+        assert(normals.length === 246, 'Cone.normal.count : ' + normals.length);
+        assert(texCoords.length === 164, 'Cone.textcoord.count : ' + texCoords.length);
         geometry.generateUniqueVertex();
         normals = geometry.attributes.normal.value;
         texCoords = geometry.attributes.texcoord0.value;
-        assert(normals.length  === 1800, 'Cylinder.normal.count : ' + normals.length);
-        assert(texCoords.length === 1200, 'Cylinder.textcoord.count : ' + texCoords.length);
+        assert(normals.length  === 720, 'Cone.normal.count : ' + normals.length);
+        assert(texCoords.length === 480, 'Cone.textcoord.count : ' + texCoords.length);
     });
 
     it('#generateBarycentric', function () {
-        const geometry = new qtek.geometry.Cylinder();
+        const geometry = new qtek.geometry.Cone();
         assert(!geometry.attributes.barycentric.value);
         geometry.generateBarycentric();
         const attrCount = geometry.attributes.barycentric.value.length;
-        assert(attrCount === 1800, attrCount);
+        assert(attrCount === 720, attrCount);
     });
 });
