+++ test/unit/geometry/cube.spec.js
@@ -1,10 +1,10 @@
-require('../common/');
+require('../../common/');
 const assert = require('assert');
-const qtek = require('../../dist/qtek');
+const qtek = require('../../../dist/qtek');
 
-describe('Plane.Spec', function () {
+describe('Cube.Spec', function () {
     it('constructor', function () {
-        const geometry = new qtek.geometry.Plane({
+        const geometry = new qtek.geometry.Cube({
             widthSegments : 1,
             heightSegments : 2,
             depthSegments : 3
@@ -19,40 +19,34 @@ describe('Plane.Spec', function () {
     });
 
     it('#generateFaceNormals', function () {
-        const geometry = new qtek.geometry.Plane();
+        const geometry = new qtek.geometry.Cube();
         const normal = geometry.attributes.normal.value;
         geometry.generateFaceNormals();
         assert.notDeepEqual(geometry.attributes.normal.value, normal);
     });
     
     it('#generateTangents', function () {
-        const geometry = new qtek.geometry.Plane();
+        const geometry = new qtek.geometry.Cube();
         const tangent = geometry.attributes.tangent.value;
         geometry.generateTangents();
         assert.notDeepEqual(geometry.attributes.tangent.value, tangent);
-        const len = geometry.attributes.tangent.value.length;
-        assert(len === 16, 'Plane.tangent : ' + len);
+        assert(geometry.attributes.tangent.value.length === 96);
     });
 
     it('#generateUniqueVertex', function () {
-        const geometry = new qtek.geometry.Plane();
+        const geometry = new qtek.geometry.Cube();
         assert(!geometry.isUniqueVertex());
-        let normals = geometry.attributes.normal.value,
-            texCoords = geometry.attributes.texcoord0.value;
-        assert(normals.length === 12, 'Plane.normal.count : ' + normals.length);
-        assert(texCoords.length === 8, 'Plane.textcoord.count : ' + texCoords.length);
+        assert(geometry.attributes.normal.value.length === 72);
+        assert(geometry.attributes.texcoord0.value.length === 48);
         geometry.generateUniqueVertex();
-        normals = geometry.attributes.normal.value;
-        texCoords = geometry.attributes.texcoord0.value;
-        assert(normals.length  === 18, 'Plane.normal.count : ' + normals.length);
-        assert(texCoords.length === 12, 'Plane.textcoord.count : ' + texCoords.length);
+        assert(geometry.attributes.normal.value.length  === 108);
+        assert(geometry.attributes.texcoord0.value.length === 72);
     });
 
     it('#generateBarycentric', function () {
-        const geometry = new qtek.geometry.Plane();
+        const geometry = new qtek.geometry.Cube();
         assert(!geometry.attributes.barycentric.value);
         geometry.generateBarycentric();
-        const attrCount = geometry.attributes.barycentric.value.length;
-        assert(attrCount === 18, attrCount);
+        assert(geometry.attributes.barycentric.value.length === 108);
     });
 });
